+++ clouddriver-openstack/src/main/groovy/com/netflix/spinnaker/clouddriver/openstack/security/OpenstackNamedAccountCredentials.groovy
@@ -24,7 +24,11 @@ class OpenstackNamedAccountCredentials implements AccountCredentials<OpenstackCr
   final String name
   final String environment
   final String accountType
-  final String lbaasVersion
+  final String compute
+  final String identity
+  final String networking
+  final String orchestration
+  final String images
   final String master
   final String username
   @JsonIgnore
@@ -40,7 +44,11 @@ class OpenstackNamedAccountCredentials implements AccountCredentials<OpenstackCr
   OpenstackNamedAccountCredentials(String accountName,
                                    String environment,
                                    String accountType,
-                                   String lbaasVersion,
+                                   String compute,
+                                   String identity,
+                                   String networking,
+                                   String orchestration,
+                                   String images,
                                    String master,
                                    String username,
                                    String password,
@@ -49,13 +57,17 @@ class OpenstackNamedAccountCredentials implements AccountCredentials<OpenstackCr
                                    String endpoint,
                                    List<String> regions,
                                    Boolean insecure) {
-    this(accountName, environment, accountType, lbaasVersion, master, username, password, null, tenantName, domainName, endpoint, regions, insecure)
+    this(accountName, environment, accountType, compute, identity, networking, orchestration, images, master, username, password, null, tenantName, domainName, endpoint, regions, insecure)
   }
 
   OpenstackNamedAccountCredentials(String accountName,
                                    String environment,
                                    String accountType,
-                                   String lbaasVersion,
+                                   String compute,
+                                   String identity,
+                                   String networking,
+                                   String orchestration,
+                                   String images,
                                    String master,
                                    String username,
                                    String password,
@@ -68,7 +80,11 @@ class OpenstackNamedAccountCredentials implements AccountCredentials<OpenstackCr
     this.name = accountName
     this.environment = environment
     this.accountType = accountType
-    this.lbaasVersion = lbaasVersion
+    this.compute = compute
+    this.identity = identity
+    this.networking = networking
+    this.orchestration = orchestration
+    this.images = images
     this.master = master
     this.username = username
     this.password = password
